default.xml: Use branch Jethro for meta-rust
The upstream of meta-rust does not provide
branches for the releases of the Yocto Project.
Using directly the master branch may lead to
compatibility issues in future. Therefore it is
better to switch to another repository with branch
Jethro which is know to work with RVI SOTA client
and its dependencies written in the Rust
programming language.

Change-Id: Ibaf5fb84aa3db86833193cf2c3ffc6e5d55ebaa1
Signed-off-by: Leon Anavi <leon.anavi@konsulko.com>