Update meta-updater
The new revision includes changes to make meta-updater layer inclusion a
neutral operation, and inheriting sota.bbclass almost neutral operation.
If we add agl-sota to the build now it should not affect anything,
and once it's there we can add "sota" DISTRO_FEATURE to machine
templates for which sota is already implemented.

https://github.com/advancedtelematic/meta-updater/compare/bae52014b6e3a49e7adf49373238dab9f5f3b4a2...8379268522e453923514cc74ef37b9585ab4bde7

Change-Id: I4501ebc226a8c52017a24c06811ee33923c692c4
Signed-off-by: Anton Gerasimov <anton@advancedtelematic.com>