Update meta-ti and remove meta-boundary
Bump meta-ti revision to the ti-2020.00-rc3 level since final release
is still a few weeks out and there have been bunch of new commits that
it would be good to pick up for any testing in the interim.

Also, meta-boundary has a zeus branch now, but inspection of meta-agl
indicates that we no longer need it for any of the existing machine
templates, so just remove it.

Bug-AGL: SPEC-2932

Signed-off-by: Scott Murray <scott.murray@konsulko.com>
Change-Id: Id8042a81a0f648b08813637022a6404912ac9c64