Switch rcar to use upstream repos
Thanks to the work done by Renesas and members of our community,
we are able to switch to the upstream repo and use it directly.
A shout-out goes to Kurokawa-san, Ronan and Stephane for the work
on the layer!

Any remaining adaptations are being worked on in
 meta-agl/meta-agl-bsp/meta-rcar-gen3
going forward.

Bug-AGL: SPEC-3443

Signed-off-by: Jan-Simon Möller <jsmoeller@linuxfoundation.org>
Change-Id: I367547ae8da2b690341e916ebe6c0ad488bf34b6