Switch default manifests to tracking honister
Switch upstream branches to honister in the default and floating
manifests where possible to track upstream.  The revisions used in
the default manifest updates are the state of the upstream repos at
the the time of this commit, and are not indicative of BSP layers
having been tested.  Note that sandbox branches are used for the
AGL repos, as this is effectively a snapshot of a WIP for potential
future reference.

Bug-AGL: SPEC-3819

Signed-off-by: Scott Murray <scott.murray@konsulko.com>
Signed-off-by: Jan-Simon Moeller <jsmoeller@linuxfoundation.org>
Change-Id: I1d418ab2a12e0a671e4bfe75316d79c123a436e5