Switch default manifests to tracking upstream
Switch upstream branches to master in the default and floating
manifests where possible to track upstream.  The revisions used in
the default manifest updates are the state of the upstream repos at
the the time of this commit, and are not indicative of BSP layers
having been tested.

Bug-AGL: SPEC-3819

Signed-off-by: Scott Murray <scott.murray@konsulko.com>
Signed-off-by: Jan-Simon Moeller <jsmoeller@linuxfoundation.org>
Change-Id: I0fa05f5111b2aac8fe3ab329bb9cc25c31f38924