Update master branch to YP 4.0 aka kirkstone
Thanks to the continuous updates from Scott on the next branch,
we can uprev directly when kirkstone is about to be released.

In this step we do uprev the AGL-repo to what is in our 'next' branch.
Then we will update the meta-agl* layers and change the default back to master.

Bug-AGL: SPEC-4343
Signed-off-by: Jan-Simon Moeller <jsmoeller@linuxfoundation.org>
Change-Id: I52b87c21be95bafa0813f6388211073ae2b8d35b