Initial kirkstone support for SanCloud BBE
Update the meta-sancloud, meta-ti & meta-rtlwifi layers to provide
initial support for the SanCloud BBE on the Yocto Project kirkstone
release. The tracking branches for both meta-sancloud & meta-ti are also
updated to point at kirkstone.

This support is still WIP and for now is limited to headless/telematics
profiles.

Bug-AGL: SPEC-4582

Signed-off-by: Paul Barker <paul.barker@sancloud.com>
Change-Id: I20d3ad0e6d56e809be9597064834c3cc14e40e2b